requirements.txt
streamlit
google-generativeai
pypdf
python-dotenv
groq